Fixes for pr comments
Have added checks for if s3 provider endpoint is minio then exit on no
creds being supplied. Readme suggests there is functionality with ec2
where creds aren't necessarily needed, so don't want to exit for
everything.

Have added a check for the path variable, will try to ls the bucket/path
and if the path does not exist it will error and will exit script.