Added metadata to improve concourse logging info
Current s3 resource didn’t always show much information on files changed, bucket or endpoint related info, when resource was passed through jobs.

This change enables the s3 encrypt resource to output resource metadata,
including attributes like bucket, s3 endpoint, path to file in bucket,
and filenames (with respective size and last modified date with
timezone)

Currently this metadata information is only viewable in the GET of a
resource from a job on concourse, as the changes were only implemented in the assets/in script.

Signed-off-by: Sushant Kapoor <sushant.kapoor@fil.com>